refactor: Renamed FeatureManager -> DatadogManager & added new file CertificatePreviewManager.
